Update ReaperSWS.download.recipe (#27)
Updated to grab the 64 bit intel version of this - the recipe fails otherwise.

As the vendor now offers 32 and 64 bit intel versions as well as an arm64 Silicon version, it may be good to have separate recipes for each arch. I'd be happy to create some and host them in our dataJAR-recipes repo - please let me know if you'd prefer that.

Thank you!